chore: gradle way of structuring applications
