+++ .gitignore
@@ -1,5 +1,4 @@
 .env
-.gradle
 **/build/
 !**/src/**/build/
 
@@ -23,6 +22,7 @@ gradle-app.setting
 
 HELP.md
 .gradle
+.kotlin
 build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
