+++ app/src/test/java/com/shopping/inandout/routeservice/RouteServiceWrapperTest.java
@@ -1,0 +1,8 @@
+package com.shopping.inandout.routeservice;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+class RouteServiceWrapperTest {
+}
